# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Project Tracking API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/customers/{customerId}/projects": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "projectYear": {
                    "nullable": true,
                    "type": "string"
                  },
                  "contractNumber": {
                    "type": "string"
                  },
                  "contractStartDate": {
                    "type": "string"
                  },
                  "contractEndDate": {
                    "type": "string"
                  },
                  "contractNotifyMonths": {
                    "type": "number"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "customerId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/customers/{customerId}/projects/{projectId}": {
      "patch": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "isActive": {
                    "type": "boolean"
                  },
                  "projectYear": {
                    "nullable": true,
                    "type": "string"
                  },
                  "contractNumber": {
                    "nullable": true,
                    "type": "string"
                  },
                  "contractStartDate": {
                    "nullable": true,
                    "type": "string"
                  },
                  "contractEndDate": {
                    "nullable": true,
                    "type": "string"
                  },
                  "contractNotifyMonths": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "customerId",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "projectId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "customerId",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "projectId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/customers/{customerId}/projects/{projectId}/duplicate": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "projectYear": {
                    "nullable": true,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "customerId",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "projectId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  }
}